+++ lib/raygun/client.rb
@@ -44,11 +44,15 @@ module Raygun
       end
 
       def error_details(exception)
-        {
+        details = {
           className:  exception.class.to_s,
           message:    exception.message.to_s.encode('UTF-16', :undef => :replace, :invalid => :replace).encode('UTF-8'),
-          stackTrace: (exception.backtrace || []).map { |line| stack_trace_for(line) }
+          stackTrace: (exception.backtrace || []).map { |line| stack_trace_for(line) },
         }
+
+        details.update(innerError: error_details(exception.cause)) if exception.respond_to?(:cause) && exception.cause
+
+        details
       end
 
       def stack_trace_for(line)